\name{antibio}
\alias{antibio}
\docType{data}
\title{Antibiotics against Shipping Fever in Calves}
\description{Hypothetical drug trial to compare the effect of four antibiotics against Shipping fever in calves 
  (Shoukri and Pause, 1999, Table 3.11).}
\usage{data(antibio)}
\format{A data frame with 24 observations on the following 3 variables.
  \describe{
    \item{treatment}{A factor with levels \code{1}, \code{2}, \code{3} and \code{4}}
    \item{n}{A numeric vector: the number of treated animals within a two-week period.}
    \item{y}{A numeric vector: the number of deaths at the end of the two weeks.}
    }
  }
\references{Shoukri, M.M., Pause, C.A., 1999, 2nd ed. \emph{Statistical methods for health sciences}. CRC Press, London.}
\keyword{datasets}
