\name{coef-methods}
\docType{methods}
\alias{coef-methods}
\alias{coef,ANY-method}
\alias{coef,glimML-method}
\alias{coef,glimQL-method}
\title{Methods for Function "coef" in Package "aod"}

\description{Extract the fixed-effect coefficients from fitted objects.}

\section{Methods}{\describe{
  \item{ANY}{Generic function: see \code{\link[stats]{coef}}.}
  \item{glimML}{Extract the estimated fixed-effect coefficients from objects of formal class \dQuote{glimML}. 
    Presently, these objects are generated by functions \code{betabin} and \code{negbin}.} 
  \item{glimQL}{Extract the estimated fixed-effect coefficients from objects of formal class \dQuote{glimQL}. 
    Presently, these objects are generated by functions \code{quasibin} and \code{quasipois}.} 
  }}

\keyword{methods}
