\name{deviance-methods}
\docType{methods}
\alias{deviance-methods}
\alias{deviance,ANY-method}
\alias{deviance,glimML-method}

\title{Methods for Function "deviance" in Package "aod"}

\description{Extracts the deviance fitted models.}

\section{Methods}{\describe{
  \item{ANY}{Generic function: see \code{\link[stats]{deviance}}.}
  \item{glimML}{Extracts the deviance from models fitted by \code{betabin} or \code{negbin}.}
  }}

\keyword{methods}
