\name{df.residual-methods}
\docType{methods}
\alias{df.residual-methods}
\alias{df.residual,ANY-method}
\alias{df.residual,glimML-method}

\title{Methods for Function "df.residual" in Package "aod"}

\description{Computes the number of degrees of freedom of the residuals from fitted objects.}

\section{Methods}{\describe{
  \item{object = "ANY"}{Generic function: see \code{\link[stats]{df.residual}}.}
  \item{object = "glimML"}{Computes the df of residuals for models fitted by \code{betabin} or \code{negbin}.}
  }}

\keyword{methods}
